Create build-test.yml
Example build workflow for use with github to overcome the armv7 build error. 
The workflow requires at least the dockerhub-repo name as secret ( see : fusionauth-containers/settings/secrets/actions ), and the fusionauth version.
The username and token secret are optional secrets if you want to push to dockerhub. (un-comment the Login to Dockerhub )